+++ examples/sync/service/outputs.tf
@@ -1,0 +1,4 @@
+output "service" {
+  description = "The Generated Sync service"
+  value       = twilio_sync_service.service
+}