+++ twilio/common/twilio_client.go
@@ -11,6 +11,7 @@ import (
 	proxy "github.com/RJPearson94/twilio-sdk-go/service/proxy/v1"
 	serverless "github.com/RJPearson94/twilio-sdk-go/service/serverless/v1"
 	studio "github.com/RJPearson94/twilio-sdk-go/service/studio/v2"
+	sync "github.com/RJPearson94/twilio-sdk-go/service/sync/v1"
 	taskrouter "github.com/RJPearson94/twilio-sdk-go/service/taskrouter/v1"
 	trunking "github.com/RJPearson94/twilio-sdk-go/service/trunking/v1"
 	verify "github.com/RJPearson94/twilio-sdk-go/service/verify/v2"
@@ -32,6 +33,7 @@ type TwilioClient struct {
 	Serverless    *serverless.Serverless
 	SIPTrunking   *trunking.Trunking
 	Studio        *studio.Studio
+	Sync          *sync.Sync
 	TaskRouter    *taskrouter.TaskRouter
 	Verify        *verify.Verify
 	Video         *video.Video